Restructure file module into multiple files
