+++ src/core/blob.ts
@@ -1,4 +1,3 @@
-
 import { a, body } from '@butility/dom/html';
 import type { VerifyBlobOptions } from '@/types/media';
 
@@ -59,8 +58,8 @@ export function downloadBlob(
 ): void {
     const contentType = blob.type || fallbackContentType;
     const url = URL.createObjectURL(new Blob([blob], { type: contentType }));
-    const link = a({ href: url, download: filename })
-    body(link)
+    const link = a({ href: url, download: filename });
+    body(link);
     link.click();
     document.body.removeChild(link);
     URL.revokeObjectURL(url);
