+++ src/core/file/zip.ts
@@ -1,81 +1,5 @@
-
-import { Request } from '@butility/network';
-import { downloadBlob } from '@/core/blob';
+import { downloadBlob } from '../blob';
 import type { ZipData, CentralDirectory } from '@/types/media';
-
-export function validateFileType(
-    file: File,
-    allowedTypes: string[],
-    allowedExtensions: string[],
-): boolean {
-    const fileType = file.type;
-    const fileName = file.name;
-    const fileExtension = fileName
-        .substring(fileName.lastIndexOf('.') + 1)
-        .toLowerCase(); // Extract extension
-
-    const mimeTypeValid = allowedTypes.some((typePattern) => {
-        const regex = new RegExp(`^${typePattern.replace('*', '.*')}$`); // Convert wildcard to regex
-        return regex.test(fileType);
-    });
-
-    const extensionValid = allowedExtensions.includes(fileExtension);
-
-    if (!mimeTypeValid || !extensionValid) {
-        console.error(
-            `File type or extension not allowed. Got type: ${fileType}, extension: ${fileExtension}`,
-        );
-        return false;
-    }
-
-    return true;
-}
-
-export function validateFileSize(file: File, maxSize: number): boolean {
-    return file.size <= maxSize;
-}
-
-export function uploadFile(file: File, url: string): Promise<any> {
-    return new Promise((resolve, reject) => {
-        const formData = new FormData();
-        formData.append('file', file);
-
-        Request.post(url, formData, {
-            useFetch: true,
-            success: (response: any) => {
-                if (response) {
-                    resolve(response);
-                } else {
-                    reject(new Error('File upload failed'));
-                }
-            }
-        });
-    });
-}
-
-export async function downloadFile(
-    fileUrl: string,
-    fileName: string,
-    fileType?: string,
-): Promise<any> {
-    return new Promise((resolve, reject) => {
-        Request.get(fileUrl, {}, {
-            useFetch: true,
-            success: (response: any) => {
-                if (response) {
-                    const blob = new Blob(
-                        [response],
-                        fileType ? { type: fileType } : {},
-                    );
-                    downloadBlob(blob, fileName);
-                    resolve(response);
-                } else {
-                    reject(new Error(`Failed to download file.`));
-                }
-            }
-        });
-    });
-}
 
 export async function unzip(zipBlob: Blob): Promise<File[]> {
     return new Promise((resolve, reject) => {
