+++ src/modules/audiovisual.ts
@@ -8,21 +8,13 @@ import {
     setCaptionTrack,
     toggleCaptions,
 } from '@/core/audiovisual/caption';
-import { 
-    controlPlayback,
-    controlVolume,
-} from "@/core/audiovisual/controls";
-import {
-    skipWithDir,
-    seekTo,
-} from "@/core/audiovisual/seek";
-import { 
-    registerVideoEvents, 
-} from "@/core/audiovisual/event"; 
+import { controlPlayback, controlVolume } from '@/core/audiovisual/controls';
+import { skipWithDir, seekTo } from '@/core/audiovisual/seek';
+import { registerVideoEvents } from '@/core/audiovisual/event';
 import {
     updateVideoSourceAndResolution,
     videoTones,
-} from "@/core/audiovisual/edit";
+} from '@/core/audiovisual/edit';
 
 const Audiovisual = {
     controlPlayback,
