+++ src/modules/file.ts
@@ -1,13 +1,9 @@
-export * from '@/core/file';
-import {
-    decompressDEFLATE,
-    downloadFile,
-    unzip,
-    uploadFile,
-    validateFileSize,
-    validateFileType,
-    zip,
-} from '@/core/file';
+export * from '@/core/file/load';
+export * from '@/core/file/validate';
+export * from '@/core/file/zip';
+import { decompressDEFLATE, unzip, zip } from '@/core/file/zip';
+import { uploadFile, downloadFile } from '@/core/file/load';
+import { validateFileSize, validateFileType } from '@/core/file/validate';
 
 const File = {
     decompressDEFLATE,
